test: validate rwx page creation with mmap
This was a big rabbit hole... I was running into the issue of attaching
early to a program you've spawned yourself versus attached to at a later
point. Attaching too early lands you in areas before the program has
initialised, so you need to wait it out a bit till it's initialised.

The ptrace guide from strace is awesome, was a huge help to understand
what was going on:
https://github.com/strace/strace/blob/master/doc/README-linux-ptrace

Still need to handle signal-delivery-stop in other areas... And
dynamically discover the entry point.