test: loading a shared library
Not sure if it's something I'm doing or a bug in Zig, but sometimes
tests seem to wait until the test exe finishes (which it won't, it's in
an infinite while loop), and gets stuck. Running the same code in main()
has no issues, it only happen when its a test.